server/requirements.txt
# Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Transcription (Groq cloud API)
httpx>=0.27.0

# Analysis (Claude)
anthropic>=0.40.0

# Database
asyncpg>=0.29.0

# Config
python-dotenv>=1.0.0
